requirements.txt
# Core dependencies
boto3>=1.34.0
PyYAML>=6.0.1

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Type stubs
boto3-stubs[essential]>=1.34.0
types-PyYAML>=6.0.12
